ActsAsDashboard::ClassMethods#acts_as_dashboard creates routes at run-time.
Routes for dashboards don't need to be defined in routes.rb anymore.

I'm not sure if this is ideal. The advantage is that users don't need to
hack routes.rb so that their dashboard works. The disadvantage is that
users who expect their dashboard-related route(s) to be in routes.rb
will be confused, and probably unable to figure out where and how the
routes are being defined.